Accept JWKs with RSA certificate if they also provide modulus and exponent as parameters
Some JWKs contain the 'n' and 'e' parameter in addition to the certificate parameters.
In this case (according to 4.6. of [RFC7517]) the values are semantically equivalent to the ones in the certificate.
Thus all information for the public key are available and no exception needs to be thrown in this situation.